Integrate the keysym into the modfield
In order to make the following work:

    @Control_L
        echo hello

we must subtract the modfield of the keysym from the modifier state of
key release event. The reason being that the modifier state we receive
is generated before the key release. It means that releasing `Control_L`
will generate a key release event whose state will most probably contain
the bit mask of the control modifier.