Provide a notation for synchronous execution
If a command starts with a semicolon it will be executed synchronously.

We need this for the *bspwm* pointer interactions: if we send the
grabbing message asynchronously, *bspwm* might not see the window below
the pointer because the pointer might have moved outside of the window
while the message was being transmitted.