Added macros handling functions
Macros are identified by a key (string to be replaced) and a replacement
text. For example, a macro defining ANIMAL to be "white cat" may be
used to process the following text "I have a {{ANIMAL}}" to produce the
string: "I have a white cat"

Note that macros can be defined using previously defined macros and
expansion would work as expected.

For example having defined COLOR to be "white", one can define ANIMAL to
be "{{COLOR}} cat", and then have the string "I have a {{ANIMAL}}" be
processed to give us "I have a white cat".